moved parseStatement to utils
